+++ src/test/suite/commands/rectangle.test.ts
@@ -212,3 +212,62 @@ KLMNOPQRST`
     assertCursorsEqual(activeTextEditor, [2, 5], [3, 9]);
   });
 });
+
+suite("open-rectangle", () => {
+  let activeTextEditor: vscode.TextEditor;
+  let emulator: EmacsEmulator;
+
+  const initialText = `0123456789
+abcdefghij
+ABCDEFGHIJ
+klmnopqrst
+KLMNOPQRST`;
+  setup(async () => {
+    activeTextEditor = await setupWorkspace(initialText);
+    emulator = new EmacsEmulator(activeTextEditor);
+  });
+
+  teardown(cleanUpWorkspace);
+
+  test("nothing happens when the selection is empty", async () => {
+    setEmptyCursors(activeTextEditor, [1, 5]);
+    await emulator.runCommand("openRectangle");
+    assertTextEqual(activeTextEditor, initialText);
+    assertCursorsEqual(activeTextEditor, [1, 5]);
+  });
+
+  test("nothing happens when the selections are empty", async () => {
+    setEmptyCursors(activeTextEditor, [1, 5], [2, 7]);
+    await emulator.runCommand("openRectangle");
+    assertTextEqual(activeTextEditor, initialText);
+    assertCursorsEqual(activeTextEditor, [1, 5], [2, 7]);
+  });
+
+  test("opening a rectangle", async () => {
+    activeTextEditor.selections = [new vscode.Selection(0, 3, 2, 7)];
+    await emulator.runCommand("openRectangle");
+    assertTextEqual(
+      activeTextEditor,
+      `012    3456789
+abc    defghij
+ABC    DEFGHIJ
+klmnopqrst
+KLMNOPQRST`
+    );
+    assertCursorsEqual(activeTextEditor, [0, 3]);
+  });
+
+  test("opening rectangles", async () => {
+    activeTextEditor.selections = [new vscode.Selection(0, 3, 2, 5), new vscode.Selection(2, 7, 3, 9)];
+    await emulator.runCommand("openRectangle");
+    assertTextEqual(
+      activeTextEditor,
+      `012  3456789
+abc  defghij
+ABC  DEFG  HIJ
+klmnopq  rst
+KLMNOPQRST`
+    );
+    assertCursorsEqual(activeTextEditor, [0, 3], [2, 7]);
+  });
+});